requirements.txt
# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.10.0
google-cloud-pubsub>=2.18.0
pandas>=2.0.0

# DBT
dbt-bigquery>=1.7.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0